feat: call .ready(err) when callback pass an error
