+++ .autod.conf.js
@@ -1,0 +1,21 @@
+'use strict'
+
+module.exports = {
+  write: true,
+  prefix: '^',
+  devprefix: '^',
+  exclude: [
+    'test/fixtures',
+  ],
+  devdep: [
+    'autod',
+    'egg-bin',
+    'egg-ci',
+    'eslint',
+    'eslint-config-egg',
+  ],
+  keep: [
+  ],
+  semver: [
+  ],
+};
