+++ .travis.yml
@@ -1,11 +1,12 @@
 sudo: false
 language: node_js
-
 node_js:
-  - 2
-  - 3
-  - 4
-
-script: npm run cov
-
-after_success: npm install codecov.io && cat ./coverage/coverage-final.json | ./node_modules/codecov.io/bin/codecov.io.js
+  - '4'
+  - '6'
+  - '7'
+install:
+  - npm i npminstall && npminstall
+script:
+  - npm run ci
+after_script:
+  - npminstall codecov && codecov
