+++ appveyor.yml
@@ -1,18 +1,16 @@
-init:
-  - git config --global core.autocrlf input
-
 environment:
   matrix:
-    - nodejs_version: 2
-    - nodejs_version: 4
+    - nodejs_version: '4'
+    - nodejs_version: '6'
+    - nodejs_version: '7'
 
 install:
   - ps: Install-Product node $env:nodejs_version
-  - npm install
-
-build: off
+  - npm i npminstall && node_modules\.bin\npminstall
 
 test_script:
   - node --version
   - npm --version
-  - npm run cov
+  - npm run ci
+
+build: off
