+++ package.json
@@ -13,19 +13,20 @@
     "lib"
   ],
   "dependencies": {
-    "debug": "~2.2.0",
-    "extend": "~3.0.0",
-    "once": "~1.3.2",
-    "get-ready": "~1.0.0",
-    "uuid": "~2.0.1"
+    "debug": "^2.6.0",
+    "get-ready": "^2.0.0",
+    "once": "^1.4.0",
+    "uuid": "^3.0.1"
   },
   "devDependencies": {
-    "istanbul": "0",
-    "eslint": "1",
-    "koa": "1",
-    "mocha": "2.3.3",
-    "should": "6",
-    "spy": "0"
+    "autod": "^2.7.1",
+    "egg-bin": "^2.0.2",
+    "egg-ci": "^1.1.0",
+    "eslint": "^3.15.0",
+    "eslint-config-egg": "^3.2.0",
+    "koa": "^1.2.4",
+    "should": "^11.2.0",
+    "spy": "^1.0.0"
   },
   "repository": {
     "type": "git",
@@ -35,12 +36,16 @@
   "author": "popomore <sakura9515@gmail.com>",
   "license": "MIT",
   "scripts": {
+    "autod": "autod",
     "lint": "eslint .",
-    "test": "mocha -R spec -t 20000",
-    "cov": "npm run lint && istanbul cover --report=none node_modules/mocha/bin/_mocha -- -R spec -t 20000 && istanbul report json html",
-    "autod": "autod -e test/fixtures -f ~ -w"
+    "test": "npm run lint -- --fix && egg-bin test",
+    "cov": "egg-bin cov",
+    "ci": "npm run lint && egg-bin cov"
   },
   "engines": {
-    "node": ">=2.0.0"
+    "node": ">=4.0.0"
+  },
+  "ci": {
+    "version": "4, 6, 7"
   }
 }
